Criando modulo do SharedPref na injeção de dependencia ajustes nos tratamentos de erros e chamadas

+++ app/src/main/java/com/facens/acedevelop/voluntariei/data/datasource/interfaces/OngDataSource.kt
@@ -9,9 +9,9 @@ interface OngDataSource {
 
     suspend fun registerOng(user: Ong): Ong
 
-    suspend fun getOng(id:Int): Ong
+    suspend fun getOng(id:Long): Ong
 
-    suspend fun deleteOng(id: Int): Boolean
+    suspend fun deleteOng(id: Long): Boolean
 
     suspend fun login(login:LoginRequest): Ong?
 }
