+++ app/src/main/java/com/facens/acedevelop/voluntariei/data/datasource/interfaces/UserDataSource.kt
@@ -7,9 +7,9 @@ interface UserDataSource {
 
     suspend fun registerUser(user:User):User
 
-    suspend fun getUser(id:Int):User
+    suspend fun getUser(id:Long):User
 
-    suspend fun deleteUser(id: Int):Boolean
+    suspend fun deleteUser(id: Long):Boolean
 
     suspend fun login(login:LoginRequest):User?
