+++ app/src/main/java/com/facens/acedevelop/voluntariei/data/repository/OngRepository.kt
@@ -10,11 +10,11 @@ class OngRepository @Inject constructor(
 ) {
     suspend fun registerOng(user: Ong): Ong = dataSource.registerOng(user)
 
-    suspend fun getOng(id:Int): Ong = dataSource.getOng(id)
+    suspend fun getOng(id:Long): Ong = dataSource.getOng(id)
 
     suspend fun updateOng(user:Ong):Ong = dataSource.updateOng(user)
 
-    suspend fun deleteOng(id: Int): Boolean = dataSource.deleteOng(id)
+    suspend fun deleteOng(id: Long): Boolean = dataSource.deleteOng(id)
 
     suspend fun loginOng(login: LoginRequest):Ong? = dataSource.login(login)
 }
