+++ app/src/main/java/com/facens/acedevelop/voluntariei/domain/models/Event.kt
@@ -5,9 +5,9 @@ import java.io.Serializable
 import java.util.*
 
 data class Event(
-    @SerializedName("id")val id:Int,
+    @SerializedName("id")val id:Long,
     @SerializedName("nome")val nome:String,
     @SerializedName("descricao")val descricao:String,
     @SerializedName("data")val data: Date,
-    @SerializedName("ong_id")val ong: Int,
+    @SerializedName("ong_id")val ong: Long,
 ) :Serializable
